requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
pillow==11.3.0
pip==25.1.1
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.2
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
typing-extensions==4.14.1
typing-inspection==0.4.1
uvicorn==0.35.0